ml/requirements.txt
imagehash>=4.3.1
Pillow>=10.0.0
httpx>=0.27.0
numpy>=1.26.0
opencv-python>=4.9.0
